refactor: Extract common helper functions and watcher utilities
タブ関連ヘルパー関数とディレクトリ監視処理を共通化。

- app/explorer_helpers.rs: タブのインデックス取得・状態更新ヘルパー
- utils/watcher_utils.rs: ExplorerとViewer共通のwatcher処理
  * subscribe_directory/unsubscribe_directory: 共通監視管理
  * create_explorer_watcher_callback: Explorer用コールバック生成
  * create_viewer_watcher_callback: Viewer用コールバック生成
- explorer.rsとviewer.rsのwatcher処理を共通関数化して重複削除
