+++ src-tauri/src/utils/mod.rs
@@ -1,1 +1,2 @@
 pub mod file_utils;
+pub mod watcher_utils;
